Folder structure for practice node server
