+++ Practice/nodeServer/package.json
@@ -13,7 +13,9 @@
   "dependencies": {
     "@types/express": "^4.17.17",
     "cors": "^2.8.5",
+    "dotenv": "^16.3.1",
     "express": "^4.18.2",
+    "googleapis": "^120.0.0",
     "jsonata": "^2.0.3"
   },
   "devDependencies": {
