+++ Practice/nodeServer/src/routers/index.ts
@@ -1,0 +1,8 @@
+import express, { Request, Response, json } from "express";
+const router = express.Router();
+
+import jsonATA from "../jsonATA/jsonATA";
+
+router.use("/jsonATA", jsonATA);
+
+export default router;
