+++ Practice/nodeServer/tsconfig.json
@@ -4,6 +4,8 @@
     "module": "commonjs",
     "outDir": "dist",
     "strict": true,
-    "esModuleInterop": true
+    "esModuleInterop": true,
+    "types": ["node", "express", "cors"],
+    "skipLibCheck": true
   }
 }
