+++ Practice/react-adv/tsconfig.json
@@ -14,8 +14,7 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "noEmit": true,
-    "jsx": "react-jsx",
-    "types": ["node", "express", "cors"]
+    "jsx": "react-jsx"
   },
   "include": ["src"]
 }
