+++ server/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "tsc-watch --noClear -p ./tsconfig.json --onSuccess \"node ./dist/index.js\"",
     "build": "tsc",
-    "migrate":"tsc && node dist/migrate.js",
+    "migrate": "tsc && node dist/migrate.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
@@ -19,6 +19,7 @@
     "express": "^4.18.2",
     "express-async-errors": "^3.1.1",
     "google-auth-library": "^8.8.0",
+    "googleapis": "^120.0.0",
     "jsonwebtoken": "^9.0.0",
     "mysql2": "^3.4.0",
     "sequelize": "^6.32.1",
