+++ server/src/index.ts
@@ -15,7 +15,7 @@ import errorHandler from "./middlewares/error.middleware";
 
 app.use(cors());
 app.use("/api", router);
-app.use(errorHandler)
+app.use(errorHandler);
 
 const port = process.env.PORT || 5000;
 app.listen(port, () => {
